1. Changed the Position of IJSONService and JSONService code file in C# for root_VS2015  2. Defined the namespace in C# explicitly to have the same service name in C# for root_VS2015
