+++ root_VS2015/programs/C#/Samples/WebApp_sample/ProjectX_sample/WebService/JSONService.svc
@@ -1,1 +1,1 @@
-﻿<%@ ServiceHost Language="C#" Debug="true" Service="JSONService"  CodeBehind="~/AppCode/JSONService.svc.cs" %>
+﻿<%@ ServiceHost Language="C#" Debug="true" Service="ProjectX_sample.JSONService"  CodeBehind="~/WebService/JSONService.svc.cs" %>
