+++ root_VS2015/programs/C#/Samples/WebApp_sample/ProjectX_sample/web.config
@@ -223,8 +223,8 @@
   </system.web>
   <system.serviceModel>
     <services>
-      <service name="JSONService">
-        <endpoint address="" binding="webHttpBinding" behaviorConfiguration="MyBehavior" contract="IJSONService"/>
+      <service name="ProjectX_sample.JSONService">
+        <endpoint address="" binding="webHttpBinding" behaviorConfiguration="MyBehavior" contract="ProjectX_sample.IJSONService"/>
       </service>
     </services>
     <behaviors>
